console changes

+++ src/de/titanium/enterprise/Entity/LivingEntity.java
@@ -148,7 +148,7 @@ public abstract class LivingEntity extends Entity {
      */
     public abstract double calculateDamage(LivingEntity enemy, int comboResult);
 
-    public abstract double calculateDefense(LivingEntity enemy, int defenseScore);
+    public abstract double calculateDefense(LivingEntity enemy, double defenseScore);
 
     /**
      * Setzt das maximale Leben des Entitys.
